move ContractFactory class
