+++ packages/client/src/contracts/ERC1271ContractFacade.ts
@@ -1,11 +1,10 @@
-import { ContractFactory } from '../ContractFactory'
-import { BrandedString, EthereumAddress, MapWithTtl, toEthereumAddress } from '@streamr/utils'
-import ERC1271ContractArtifact from '../ethereumArtifacts/IERC1271Abi.json'
-import type { IERC1271 as ERC1271Contract } from '../ethereumArtifacts/IERC1271'
-import { Mapping } from '../utils/Mapping'
+import { BrandedString, EthereumAddress, MapWithTtl, hash, recoverAddress, toEthereumAddress } from '@streamr/utils'
 import { Lifecycle, scoped } from 'tsyringe'
-import { recoverAddress, hash } from '@streamr/utils'
 import { RpcProviderSource } from '../RpcProviderSource'
+import type { IERC1271 as ERC1271Contract } from '../ethereumArtifacts/IERC1271'
+import ERC1271ContractArtifact from '../ethereumArtifacts/IERC1271Abi.json'
+import { Mapping } from '../utils/Mapping'
+import { ContractFactory } from './ContractFactory'
 
 export const SUCCESS_MAGIC_VALUE = '0x1626ba7e' // Magic value for success as defined by ERC-1271
