+++ packages/client/src/contracts/StorageNodeRegistry.ts
@@ -2,13 +2,13 @@ import { EthereumAddress, toEthereumAddress } from '@streamr/utils'
 import { Lifecycle, inject, scoped } from 'tsyringe'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
 import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
-import { ContractFactory } from '../ContractFactory'
-import { getEthersOverrides } from '../ethereumUtils'
+import { RpcProviderSource } from '../RpcProviderSource'
 import { StreamrClientError } from '../StreamrClientError'
 import type { NodeRegistry as NodeRegistryContract } from '../ethereumArtifacts/NodeRegistry'
 import NodeRegistryArtifact from '../ethereumArtifacts/NodeRegistryAbi.json'
+import { getEthersOverrides } from '../ethereumUtils'
+import { ContractFactory } from './ContractFactory'
 import { waitForTx } from './contract'
-import { RpcProviderSource } from '../RpcProviderSource'
 
 export interface StorageNodeMetadata {
     urls: string[]
