+++ packages/client/src/contracts/StreamRegistry.ts
@@ -4,7 +4,6 @@ import { ContractTransactionResponse } from 'ethers'
 import { Lifecycle, inject, scoped } from 'tsyringe'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
 import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
-import { ContractFactory } from '../ContractFactory'
 import { RpcProviderSource } from '../RpcProviderSource'
 import { Stream, StreamMetadata } from '../Stream'
 import { StreamIDBuilder } from '../StreamIDBuilder'
@@ -29,10 +28,11 @@ import {
 import { filter, map } from '../utils/GeneratorUtils'
 import { LoggerFactory } from '../utils/LoggerFactory'
 import { CacheAsyncFn, CacheAsyncFnType } from '../utils/caches'
-import { ObservableContract, initContractEventGateway, waitForTx } from './contract'
 import { until } from '../utils/promises'
 import { StreamFactory } from './../StreamFactory'
 import { ChainEventPoller } from './ChainEventPoller'
+import { ContractFactory } from './ContractFactory'
+import { ObservableContract, initContractEventGateway, waitForTx } from './contract'
 import { SearchStreamsOrderBy, SearchStreamsPermissionFilter, searchStreams as _searchStreams } from './searchStreams'
 
 /*
